Take nav header out of container
Because the container is a flex container, and this visually hidden element was seen as a flex item in some browsers (IE11). Taking it out of the container div doesn't dramatically change the document structure and fixes this, as now no browser can accidentically count it into the container's flex measurements.

Signed-off-by: Hidde de Vries <hidde@hiddedevries.nl>